move card to it's own component
