+++ src/views/Analysis/Overview/AnalysisOverview.tsx
@@ -1,42 +1,15 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { AppState, FinishedGameState } from '../../../state/AppState';
-import { BigBoard } from '../../../components/Board/BigBoard/BigBoard';
-import {
-  Card,
-  CardPrimaryAction,
-  CardActions,
-  CardActionButton,
-} from '@rmwc/card';
 import { Typography } from '@rmwc/typography';
 import { ListDivider } from '@rmwc/list';
-import { Icon } from '@rmwc/icon';
 import styles from './AnalysisOverview.module.css';
 import appRoutes from '../../../routes/routes';
 import { Link } from 'react-router-dom';
-import { formatDistanceStrict } from 'date-fns';
+import { GameSummaryCard } from '../../../components/Analysis/GameSummaryCard/GameSummaryCard';
 
 interface AnalysisOverviewProps {
   finishedGames: FinishedGameState[];
-}
-
-function getGameSummary(winner: 'X' | 'O' | null, moves: number) {
-  if (winner) {
-    return (
-      <>
-        <Icon
-          icon={{
-            icon: winner.toLowerCase(),
-            size: 'medium',
-          }}
-          className={styles.winnerIcon}
-        />
-        won after {moves} moves.
-      </>
-    );
-  }
-
-  return <>{moves} moves resulted in draw.</>;
 }
 
 function getLink(game: FinishedGameState) {
@@ -52,9 +25,6 @@ function getLink(game: FinishedGameState) {
 export function AnalysisOverview(props: AnalysisOverviewProps) {
   const { finishedGames } = props;
 
-  // TODO: don't animate the symbols on the bigBoard
-  // TODO: Break this out into multiple components
-
   return (
     <div className="centerHorizontal">
       <div className={styles.analysisOverviewLayout}>
@@ -66,49 +36,12 @@ export function AnalysisOverview(props: AnalysisOverviewProps) {
         <ListDivider />
         <div className={styles.gameList}>
           {finishedGames.map((game, index) => (
-            <Card key={game.id ? game.id : game.date}>
-              <CardPrimaryAction style={{ padding: '10px' }} {...getLink(game)}>
-                <div style={{ padding: '10px' }}>
-                  <div>
-                    <Typography use={'headline4'}>
-                      Game No. {finishedGames.length - index}
-                    </Typography>
-                  </div>
-                  <div>
-                    <Typography use={'subtitle1'}>
-                      {getGameSummary(game.winner, game.moves.length)}
-                    </Typography>
-                  </div>
-                  <div style={{ paddingBottom: '10px' }}>
-                    <Typography
-                      use={'subtitle2'}
-                      style={{ fontStyle: 'italic' }}
-                    >
-                      {formatDistanceStrict(
-                        new Date(),
-                        new Date(Date.parse(game.date)),
-                      )}{' '}
-                      ago
-                    </Typography>
-                  </div>
-                  <ListDivider />
-                </div>
-                <BigBoard
-                  currentPlayer={game.moves[game.moves.length - 2].player}
-                  board={game.gameState}
-                  activeBoards={[]}
-                  onPlayerMoved={() => {}}
-                  animate={false}
-                />
-              </CardPrimaryAction>
-              <CardActions>
-                <CardActionButton
-                  label="Analyse Game"
-                  trailingIcon="arrow-right"
-                  {...getLink(game)}
-                />
-              </CardActions>
-            </Card>
+            <GameSummaryCard
+              key={game.id ? game.id : game.date}
+              gameNumber={finishedGames.length - index}
+              game={game}
+              link={getLink(game)}
+            />
           ))}
         </div>
       </div>
